Use bigint for friend discount telegram ids

+++ app/models/friend_discount.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, Integer, String
+from sqlalchemy import BigInteger, Boolean, DateTime, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.db import Base
@@ -11,7 +11,7 @@ class FriendDiscount(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
 
-    telegram_id: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
+    telegram_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
     discount_percent: Mapped[int] = mapped_column(Integer, nullable=False)
 
     max_usages: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
